+++ Sources/Core/Payment/PaymentMethodDetails.swift
@@ -37,7 +37,6 @@ public enum PaymentMethodUserData: Codable, Equatable {
     case datatransCardAlias(DatatransCreditCardData)
     case payoneCreditCard(PayoneCreditCardData)
     case payoneSepa(PayoneSepaData)
-    case leinweberCustomerNumber(LeinweberCustomerData)
     case invoiceByLogin(InvoiceByLoginData)
 
     public enum CodingKeys: String, CodingKey {
@@ -65,7 +64,6 @@ public enum PaymentMethodUserData: Codable, Equatable {
         case .datatransCardAlias(let data): return data
         case .payoneCreditCard(let data): return data
         case .payoneSepa(let data): return data
-        case .leinweberCustomerNumber(let data): return data
         case .invoiceByLogin(let data): return data
         }
     }
@@ -98,8 +96,6 @@ public enum PaymentMethodUserData: Codable, Equatable {
             self = .payoneCreditCard(payoneData)
         } else if let payoneSepa = try container.decodeIfPresent(PayoneSepaData.self, forKey: .payoneSepa) {
             self = .payoneSepa(payoneSepa)
-       } else if let leinweberData = try container.decodeIfPresent(LeinweberCustomerData.self, forKey: .leinweberCustomerNumber) {
-            self = .leinweberCustomerNumber(leinweberData)
        } else if let invoiceData = try container.decodeIfPresent(InvoiceByLoginData.self, forKey: .invoiceByLogin) {
             self = .invoiceByLogin(invoiceData)
         } else {
@@ -118,7 +114,6 @@ public enum PaymentMethodUserData: Codable, Equatable {
         case .datatransCardAlias(let data): try container.encode(data, forKey: .datatransCardAlias)
         case .payoneCreditCard(let data): try container.encode(data, forKey: .payoneCreditCard)
         case .payoneSepa(let data): try container.encode(data, forKey: .payoneSepa)
-        case .leinweberCustomerNumber(let data): try container.encode(data, forKey: .leinweberCustomerNumber)
         case .invoiceByLogin(let data): try container.encode(data, forKey: .invoiceByLogin)
         }
     }
@@ -195,11 +190,6 @@ public struct PaymentMethodDetail: Equatable {
         self.methodData = PaymentMethodUserData.tegutEmployeeCard(tegutData)
     }
 
-    public init(_ leinweberData: LeinweberCustomerData) {
-        self.id = UUID()
-        self.methodData = PaymentMethodUserData.leinweberCustomerNumber(leinweberData)
-    }
-
     public init(_ paydirektData: GiropayData) {
         self.id = UUID()
         self.methodData = PaymentMethodUserData.giropayAuthorization(paydirektData)
@@ -256,9 +246,7 @@ public struct PaymentMethodDetail: Equatable {
     public var rawMethod: RawPaymentMethod {
         switch self.methodData {
         case .sepa, .payoneSepa: return .deDirectDebit
-        case .tegutEmployeeCard,
-                .leinweberCustomerNumber,
-                .invoiceByLogin:
+        case .tegutEmployeeCard, .invoiceByLogin:
             return .externalBilling
         case .giropayAuthorization:
             return .giropayOneKlick
@@ -296,7 +284,7 @@ public struct PaymentMethodDetail: Equatable {
             return payoneSepaData.projectId
         case .invoiceByLogin(let invoiceData):
             return invoiceData.projectId
-        case .sepa, .tegutEmployeeCard, .giropayAuthorization, .leinweberCustomerNumber:
+        case .sepa, .tegutEmployeeCard, .giropayAuthorization:
             return nil
         }
     }
@@ -540,27 +528,6 @@ extension PaymentMethodDetails {
 
     public static func removeTegutEmployeeCard() {
         let details = self.read().filter { $0.originType != .tegutEmployeeID }
-        self.save(details)
-    }
-}
-
-// extensions for leinweber customer numbers that can be used as payment methods
-extension PaymentMethodDetails {
-    public static func addLeinweberCustomerNumber(_ number: String, _ name: String, _ projectId: Identifier<Project>) {
-        guard
-            let cert = Snabble.shared.certificates.first,
-            let employeeData = LeinweberCustomerData(cert.data, number, name, projectId)
-        else {
-            return
-        }
-
-        var details = self.read().filter { $0.originType != .leinweberCustomerID }
-        details.append(PaymentMethodDetail(employeeData))
-        self.save(details)
-    }
-
-    public static func removeLeinweberCustomerNumber() {
-        let details = self.read().filter { $0.originType != .leinweberCustomerID }
         self.save(details)
     }
 }